Renamed SkypeMiddleware to SkypMentionNormalizeMiddleware, added more comments, and simplified the Mention_Skype test.
