+++ libraries/Microsoft.Bot.Schema/ActivityExtensions.cs
@@ -44,21 +44,23 @@ namespace Microsoft.Bot.Schema
         }
 
         /// <summary>
-        /// Remove any mention text for given id from Text property.  For example, given the message
+        /// Remove any mention text for given id from the Activity.Text property.  For example, given the message
         /// @echoBot Hi Bot, this will remove "@echoBot", leaving "Hi Bot".
         /// </summary>
         /// <description>
+        /// Typically this would be used to remove the mention text for the target recipient (the bot usually), though
+        /// it could be called for each member.  For example:
+        ///    turnContext.Activity.RemoveMentionText(turnContext.Activity.Recipient.Id);
         /// The format of a mention Activity.Entity is dependent on the Channel.  But in all cases we
         /// expect the Mention.Text to contain the exact text for the user as it appears in
         /// Activity.Text.
-        /// 
         /// For example, Teams uses &lt;at&gt;username&lt;/at&gt;, whereas slack use @username. It
         /// is expected that text is in Activity.Text and this method will remove that value from
         /// Activity.Text.
         /// </description>
-        /// <param name="activity">activity.</param>
+        /// <param name="activity">activity.  The Text property is modified by this method.</param>
         /// <param name="id">id to match.</param>
-        /// <returns>new .Text property value.</returns>
+        /// <returns>new Activity.Text property value.</returns>
         public static string RemoveMentionText(this IMessageActivity activity, string id)
         {
             foreach (var mention in activity.GetMentions().Where(mention => mention.Mentioned.Id == id))
